+++ modules/resource-metadata-sqs/main.tf
@@ -14,10 +14,16 @@ locals {
     Provider = "Coralogix"
     License  = "Apache-2.0"
   }
+
+  sns_kms_key_resource = coalesce(
+    var.sns_kms_key_arn,
+    "arn:${data.aws_partition.current.partition}:kms:${data.aws_region.this.id}:${data.aws_caller_identity.current.account_id}:key/00000000-0000-0000-0000-000000000000"
+  )
 }
 
 data "aws_region" "this" {}
 data "aws_caller_identity" "current" {}
+data "aws_partition" "current" {}
 
 resource "aws_sqs_queue" "metadata_queue" {
   name                       = "${local.function_name}-metadata-queue"
@@ -185,7 +191,15 @@ module "collector_lambda" {
         ]
         resources = [var.crossaccount_config_assume_role]
       }
-    } : {}
+    } : {},
+    {
+      sns_kms = {
+        sid       = "SnsKms"
+        effect    = "Allow"
+        actions   = ["kms:Decrypt", "kms:GenerateDataKey*"]
+        resources = [local.sns_kms_key_resource]
+      }
+    }
   )
 
   allowed_triggers = {
@@ -273,7 +287,15 @@ module "generator_lambda" {
         ]
         resources = ["arn:aws:iam::*:role/${var.crossaccount_iam_role_name}"]
       }
-    } : {}
+    } : {},
+    {
+      sns_kms = {
+        sid       = "SnsKms"
+        effect    = "Allow"
+        actions   = ["kms:Decrypt", "kms:GenerateDataKey*"]
+        resources = [local.sns_kms_key_resource]
+      }
+    }
   )
 
   allowed_triggers = {
@@ -384,7 +406,15 @@ module "generator_lambda_sm" {
         ]
         resources = ["arn:aws:iam::*:role/${var.crossaccount_iam_role_name}"]
       }
-    } : {}
+    } : {},
+    {
+      sns_kms = {
+        sid       = "SnsKms"
+        effect    = "Allow"
+        actions   = ["kms:Decrypt", "kms:GenerateDataKey*"]
+        resources = [local.sns_kms_key_resource]
+      }
+    }
   )
 
   allowed_triggers = {
@@ -407,9 +437,10 @@ module "generator_lambda_sm" {
 }
 
 resource "aws_sns_topic" "this" {
-  name_prefix  = "${local.function_name}-Failure"
-  display_name = "${local.function_name}-Failure"
-  tags         = merge(var.tags, local.tags)
+  name_prefix       = "${local.function_name}-Failure"
+  display_name      = "${local.function_name}-Failure"
+  kms_master_key_id = var.sns_kms_key_arn
+  tags              = merge(var.tags, local.tags)
 }
 
 resource "aws_secretsmanager_secret" "api_key_secret" {